Moved all singularity-related download functions into a new file
