+++ .gitignore
@@ -18,7 +18,6 @@ env/
 build/
 develop-eggs/
 dist/
-downloads/
 eggs/
 .eggs/
 lib64/
