+++ nf_core/pipelines/downloads/utils.py
@@ -1,0 +1,41 @@
+import rich.progress
+
+
+class DownloadError(RuntimeError):
+    """A custom exception that is raised when nf-core pipelines download encounters a problem that we already took into consideration.
+    In this case, we do not want to print the traceback, but give the user some concise, helpful feedback instead.
+    """
+
+
+class DownloadProgress(rich.progress.Progress):
+    """Custom Progress bar class, allowing us to have two progress
+    bars with different columns / layouts.
+    """
+
+    def get_renderables(self):
+        for task in self.tasks:
+            if task.fields.get("progress_type") == "summary":
+                self.columns = (
+                    "[magenta]{task.description}",
+                    rich.progress.BarColumn(bar_width=None),
+                    "[progress.percentage]{task.percentage:>3.0f}%",
+                    "•",
+                    "[green]{task.completed}/{task.total} completed",
+                )
+            if task.fields.get("progress_type") == "download":
+                self.columns = (
+                    "[blue]{task.description}",
+                    rich.progress.BarColumn(bar_width=None),
+                    "[progress.percentage]{task.percentage:>3.1f}%",
+                    "•",
+                    rich.progress.DownloadColumn(),
+                    "•",
+                    rich.progress.TransferSpeedColumn(),
+                )
+            if task.fields.get("progress_type") == "singularity_pull":
+                self.columns = (
+                    "[magenta]{task.description}",
+                    "[blue]{task.fields[current_log]}",
+                    rich.progress.BarColumn(bar_width=None),
+                )
+            yield self.make_tasks_table([task])
